requirements.txt
undetected-chromedriver>=3.5.3
selenium>=4.15.2
pandas>=2.0.0
prefect>=2.14.0
python-dotenv>=1.0.0
gitpython>=3.1.0
